Refaktorerer PersonDao
- Bruker Language-dekoratøren for å få syntax-highlighting i queries
- Bruker map i stedet for spørsmålstegn for å injecte verdier i queries
- Inliner overflødige funksjoner
- Returnerer long ved insert, ikke int

+++ src/main/kotlin/no/nav/helse/mediator/meldinger/HentEnhetløsning.kt
@@ -20,8 +20,8 @@ internal class HentEnhetløsning(private val enhetNr: String) {
         personDao: PersonDao,
         fødselsnummer: String,
         aktørId: String,
-        navnId: Int,
-        infotrygdutbetalingerId: Int
+        navnId: Long,
+        infotrygdutbetalingerId: Long
     ) =
         personDao.insertPerson(
             fødselsnummer = fødselsnummer,
